feat: 충전소 상태 단순 크롤링(같은 충전소끼리 아직 묶지 않았음)

+++ crawling/crawling.py
@@ -4,11 +4,28 @@ import json
 import requests
 import schedule
 import time
+import traceback
 
-def getSecret(key, secrets):
+def writeLog(log: str):
+    current_time = time.strftime("%Y.%m.%d/%H:%M:%S", time.localtime(time.time()))
+    with open("Log.txt", "a") as f:
+        f.write(f"[{current_time}] - {log}\n")
+
+def chargerLog(str: str):
+    current_time = time.strftime("%Y.%m.%d/%H:%M:%S", time.localtime(time.time()))
+    with open("fetched_charger_log.txt", "w") as f:
+        f.write(f"[{current_time}] - {str}\n")
+            
+def getSecret(key):
     '''
     이름이 key인 비공개 key 리턴
     '''
+    # 서비스키는 secrets.json에서 관리하고 .gitignore 처리
+    BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+    secret_file = os.path.join(BASE_DIR, 'crawling/secrets.json')
+    with open(secret_file) as f:
+        secrets = json.loads(f.read())
+
     try:
         return secrets[key]
     except KeyError:
@@ -20,7 +37,7 @@ def connectDB():
     crawling DB 연결
     '''
     # mongoDB 서버에 연결
-    client = MongoClient('mongodb://minsuhan:minsuhan@13.125.236.143', 27017)
+    client = MongoClient('mongodb://minsuhan:minsuhan@localhost/?authSource=admin', 27017)
     # crawling 데이터베이스에 접근
     db = client['crawling']
     return db
@@ -34,19 +51,13 @@ def getEvChargerInfo():
     base_url = 'http://apis.data.go.kr/B552584/EvCharger/getChargerInfo'
 
     # requests.get() 함수가 parameter를 인코딩해주기 때문에 디코딩된 키를 전달해야함
-    # TODO: 서비스키는 secrets.json에서 관리하고 .gitignore 처리
-    BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-    secret_file = os.path.join(BASE_DIR, 'crawling/secrets.json')
-    with open(secret_file) as f:
-        secrets = json.loads(f.read())
-    
     # 디코딩된 serviceKey
-    serviceKey_decoded = getSecret('serviceKeyDecoded', secrets=secrets)
+    serviceKey_decoded = getSecret('serviceKeyDecoded')
 
     # TODO: 전체 충전소 검색 시 상세조건 협의 필요
     params = {
         'serviceKey': serviceKey_decoded,
-        'pageNo': '1', 'numOfRows': '10', 'zcode': '30', 'dataType': 'JSON',
+        'pageNo': '1', 'numOfRows': '9999', 'zcode': '30', 'dataType': 'JSON',
     }
 
     # fetch 후 충전소 리스트만 추출하여 리턴
@@ -55,11 +66,13 @@ def getEvChargerInfo():
         # HTTP 에러가 발생하면 예외 발생
         response.raise_for_status()  
         # 정상적으로 결과를 받은 경우
-        if (response.status_code == 200):
+        if (response.status_code == 200 and ('<errMsg>' not in response.text)):
+            # 가져온 리스트 로그 쓰기 (임시 에러확인용)
+            chargerLog(response.text)
             response = response.json()
             return response['items']['item']
         else:
-            raise Exception('HTTP Error: ' + response.status_code)
+            return False
     except requests.exceptions.RequestException as e:
         return f"Error fetching charger info: {e}"
     except Exception as e:
@@ -75,23 +88,31 @@ def updateChargers(db):
     collection = db['chargers']
     # 공공데이터로부터 충전소 리스트 가져오기
     chargers = getEvChargerInfo()
-    # 컬렉션 기존 문서 제거
-    collection.delete_many({})
-    # 컬렉션에 추가
-    print(len(chargers))
-    collection.insert_many(chargers)
-    # 추가된 시간 timestamp 추가
-    collection.update_many({}, [{"$set": { "timestamp": {"$toDate":"$_id"}}}])
-    print('Successfully Fetched and updated!')
+    if chargers is not False:
+        # 컬렉션 기존 문서 제거
+        collection.delete_many({})
+        # 컬렉션에 추가
+        collection.insert_many(chargers)
+        # 추가된 시간 timestamp 추가
+        collection.update_many({}, [{"$set": { "timestamp": {"$toDate":"$_id"}}}])
+        writeLog('Successfully Fetched and updated!')
+    else:
+        writeLog('공공데이터 서버 응답 문제가 발생하여 크롤링을 1회 skip합니다...')
 
 
 
 if __name__ == "__main__":
-    db = connectDB()
-     # schedule 라이브러리를 이용하여 일정 주기마다 job 함수 실행
-    schedule.every(5).minutes.do(updateChargers, db)
+    try:
+        db = connectDB()
+        # schedule 라이브러리를 이용하여 일정 주기마다 job 함수 실행
+        schedule.every(2).minutes.do(updateChargers, db)
 
-    # 무한 루프를 돌면서 스케줄을 체크 (필수 코드)
-    while True:
-      schedule.run_pending()
-      time.sleep(1)
+        # 무한 루프를 돌면서 스케줄을 체크 (필수 코드)
+        while True:
+            schedule.run_pending()
+            time.sleep(1)
+    except Exception:
+        err = traceback.format_exc()
+        writeLog(str('[main process Error] ' + err))
+        # 혹시나 프로그램 자체가 뻗는경우 shell에 재실행 명령
+        os.system("nohup python crawling.py &")
